requirements.txt
fastapi
uvicorn
pydantic
pandas
snowflake-connector-python
python-dotenv
pydantic[email]git add requirements.txt


